% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate Predictor-Outcome Relationship}
\usage{
calibrate(predictor, outcome)
}
\arguments{
\item{predictor}{A numeric vector representing the predictor variable.}

\item{outcome}{A numeric vector representing the outcome variable.}
}
\value{
A step function that represents the calibrated relationship between the predictor and the outcome.
}
\description{
This function calibrates the predictor-outcome relationship using isotonic regression.
}
