% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cross_validate}
\alias{cross_validate}
\title{Cross-Validate Heterogeneous Treatment Effect Models}
\usage{
cross_validate(
  hte_learners,
  hte3_task,
  cv_metalearner = Lrnr_cv_selector$new(loss_squared_error),
  cv_control = NULL,
  ...
)
}
\arguments{
\item{hte_learners}{A single \code{Lrnr_hte} learner, a list of
\code{Lrnr_hte} learners, or a \code{Stack} of \code{Lrnr_hte} learners
to cross-validate.}

\item{hte3_task}{An \code{hte3_Task} object containing the data and necessary information for heterogeneous treatment effect estimation.}

\item{cv_metalearner}{An optional metalearner (\code{Lrnr_base} object) used to combine the cross-validated learners. Default is \code{Lrnr_cv_selector$new(loss_squared_error)}.}

\item{cv_control}{A list of control parameters for cross-validation passed to \code{\link[sl3]{Lrnr_sl}}. Default is \code{NULL}.}

\item{...}{Additional arguments to pass to the loss function and other functions.}
}
\value{
A trained \code{Lrnr_sl} object containing the cross-validated ensemble of heterogeneous treatment effect learners.
}
\description{
Cross-validates a collection of heterogeneous treatment effect learners.
}
