% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hte3_Task.R
\docType{class}
\name{hte3_Task}
\alias{hte3_Task}
\title{Task object for meta-learners in causal data structures.}
\format{
\code{\link{R6Class}} object.
}
\value{
An object of class \code{hte3_Task}.
}
\description{
Constructs a \code{hte3_Task} object for meta-learners in causal data structures, containing relevant data and nuisance function estimators.
}
\details{
Class for Storing Data, NPSEM, and Nuisance Function Estimators for hte3 Learners

This class inherits from \code{\link[sl3]{sl3_Task}} and \code{\link[tmle3]{tmle3_Task}}.
In addition to all the methods supported by \code{\link[sl3]{sl3_Task}} and \code{\link[tmle3]{tmle3_Task}},
it supports the following functionalities specific to the heterogeneous treatment effect estimation (hte3) framework.
}
\keyword{data}
\section{Super classes}{
\code{\link[sl3:sl3_Task]{sl3::sl3_Task}} -> \code{\link[tmle3:tmle3_Task]{tmle3::tmle3_Task}} -> \code{hte3_Task}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-hte3_Task-new}{\code{hte3_Task$new()}}
\item \href{#method-hte3_Task-add_nuisance_estimator}{\code{hte3_Task$add_nuisance_estimator()}}
\item \href{#method-hte3_Task-get_nuisance_estimates}{\code{hte3_Task$get_nuisance_estimates()}}
\item \href{#method-hte3_Task-next_in_chain}{\code{hte3_Task$next_in_chain()}}
\item \href{#method-hte3_Task-subset_task}{\code{hte3_Task$subset_task()}}
\item \href{#method-hte3_Task-clone}{\code{hte3_Task$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sl3" data-topic="sl3_Task" data-id="add_columns"><a href='../../sl3/html/sl3_Task.html#method-sl3_Task-add_columns'><code>sl3::sl3_Task$add_columns()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="sl3_Task" data-id="add_interactions"><a href='../../sl3/html/sl3_Task.html#method-sl3_Task-add_interactions'><code>sl3::sl3_Task$add_interactions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="sl3_Task" data-id="get_data"><a href='../../sl3/html/sl3_Task.html#method-sl3_Task-get_data'><code>sl3::sl3_Task$get_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="sl3_Task" data-id="get_folds"><a href='../../sl3/html/sl3_Task.html#method-sl3_Task-get_folds'><code>sl3::sl3_Task$get_folds()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="sl3_Task" data-id="get_node"><a href='../../sl3/html/sl3_Task.html#method-sl3_Task-get_node'><code>sl3::sl3_Task$get_node()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="sl3_Task" data-id="has_node"><a href='../../sl3/html/sl3_Task.html#method-sl3_Task-has_node'><code>sl3::sl3_Task$has_node()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="sl3_Task" data-id="offset_transformed"><a href='../../sl3/html/sl3_Task.html#method-sl3_Task-offset_transformed'><code>sl3::sl3_Task$offset_transformed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="sl3_Task" data-id="revere_fold_task"><a href='../../sl3/html/sl3_Task.html#method-sl3_Task-revere_fold_task'><code>sl3::sl3_Task$revere_fold_task()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Task" data-id="generate_counterfactual_task"><a href='../../tmle3/html/tmle3_Task.html#method-tmle3_Task-generate_counterfactual_task'><code>tmle3::tmle3_Task$generate_counterfactual_task()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Task" data-id="get_node_bounds"><a href='../../tmle3/html/tmle3_Task.html#method-tmle3_Task-get_node_bounds'><code>tmle3::tmle3_Task$get_node_bounds()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Task" data-id="get_regression_task"><a href='../../tmle3/html/tmle3_Task.html#method-tmle3_Task-get_regression_task'><code>tmle3::tmle3_Task$get_regression_task()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Task" data-id="get_tmle_node"><a href='../../tmle3/html/tmle3_Task.html#method-tmle3_Task-get_tmle_node'><code>tmle3::tmle3_Task$get_tmle_node()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Task" data-id="print"><a href='../../tmle3/html/tmle3_Task.html#method-tmle3_Task-print'><code>tmle3::tmle3_Task$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Task" data-id="scale"><a href='../../tmle3/html/tmle3_Task.html#method-tmle3_Task-scale'><code>tmle3::tmle3_Task$scale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Task" data-id="unscale"><a href='../../tmle3/html/tmle3_Task.html#method-tmle3_Task-unscale'><code>tmle3::tmle3_Task$unscale()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hte3_Task-new"></a>}}
\if{latex}{\out{\hypertarget{method-hte3_Task-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hte3_Task$new(data, npsem, likelihood = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A named data frame or data.table containing treatment effect modifiers,
potential confounders, treatment, outcome, and optionally weights and subject IDs.
See the \code{\link{sl3_Task}} documentation for further options.}

\item{\code{npsem}}{A list containing \code{\link[tmle3]{tmle3_Node}} objects containing the non-parametric structural equation model (NPSEM)
specifying the relationship between variables in the treatment effect estimation framework.}

\item{\code{likelihood}}{An \code{\link[tmle3]{Likelihood}} object specifying the relevant likelihood/nuisance estimators for the meta-learner.
used for estimating the parameters of the NPSEM.}

\item{\code{...}}{Additional arguments to pass to the initialization function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hte3_Task-add_nuisance_estimator"></a>}}
\if{latex}{\out{\hypertarget{method-hte3_Task-add_nuisance_estimator}{}}}
\subsection{Method \code{add_nuisance_estimator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hte3_Task$add_nuisance_estimator(node, learner)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hte3_Task-get_nuisance_estimates"></a>}}
\if{latex}{\out{\hypertarget{method-hte3_Task-get_nuisance_estimates}{}}}
\subsection{Method \code{get_nuisance_estimates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hte3_Task$get_nuisance_estimates(
  nodes,
  hte3_task = NULL,
  fold_number = "validation"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hte3_Task-next_in_chain"></a>}}
\if{latex}{\out{\hypertarget{method-hte3_Task-next_in_chain}{}}}
\subsection{Method \code{next_in_chain()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hte3_Task$next_in_chain(
  covariates = NULL,
  outcome = NULL,
  id = NULL,
  weights = NULL,
  offset = NULL,
  time = NULL,
  folds = NULL,
  column_names = NULL,
  new_nodes = NULL,
  new_outcome_type = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments to pass to the initialization function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hte3_Task-subset_task"></a>}}
\if{latex}{\out{\hypertarget{method-hte3_Task-subset_task}{}}}
\subsection{Method \code{subset_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hte3_Task$subset_task(row_index, drop_folds = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hte3_Task-clone"></a>}}
\if{latex}{\out{\hypertarget{method-hte3_Task-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hte3_Task$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
