% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_cate_EP.R
\name{Lrnr_cate_EP}
\alias{Lrnr_cate_EP}
\title{Lrnr_cate_EP Class}
\format{
An R6 class with public methods to initialize the learner, create a regression task, and access the base learner.
}
\description{
This class defines the EP-learner of van der Laan et al. (2023) for estimation of the conditional average treatment effect.
}
\details{
The EP-learner is a robust and doubly-robust meta-learner that inherits desirable properties of both T-learner and DR-learner.
}
\section{Super classes}{
\code{\link[sl3:Lrnr_base]{sl3::Lrnr_base}} -> \code{hte3::Lrnr_hte} -> \code{Lrnr_cate_EP}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Lrnr_cate_EP-new}{\code{Lrnr_cate_EP$new()}}
\item \href{#method-Lrnr_cate_EP-get_pseudo_data}{\code{Lrnr_cate_EP$get_pseudo_data()}}
\item \href{#method-Lrnr_cate_EP-clone}{\code{Lrnr_cate_EP$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="assert_trained"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-assert_trained'><code>sl3::Lrnr_base$assert_trained()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="base_chain"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-base_chain'><code>sl3::Lrnr_base$base_chain()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="base_predict"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-base_predict'><code>sl3::Lrnr_base$base_predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="base_train"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-base_train'><code>sl3::Lrnr_base$base_train()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="chain"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-chain'><code>sl3::Lrnr_base$chain()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="custom_chain"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-custom_chain'><code>sl3::Lrnr_base$custom_chain()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="get_outcome_range"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-get_outcome_range'><code>sl3::Lrnr_base$get_outcome_range()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="get_outcome_type"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-get_outcome_type'><code>sl3::Lrnr_base$get_outcome_type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="predict"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-predict'><code>sl3::Lrnr_base$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="predict_fold"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-predict_fold'><code>sl3::Lrnr_base$predict_fold()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="print"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-print'><code>sl3::Lrnr_base$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="process_formula"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-process_formula'><code>sl3::Lrnr_base$process_formula()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="reparameterize"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-reparameterize'><code>sl3::Lrnr_base$reparameterize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="retrain"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-retrain'><code>sl3::Lrnr_base$retrain()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="sample"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-sample'><code>sl3::Lrnr_base$sample()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="set_train"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-set_train'><code>sl3::Lrnr_base$set_train()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="subset_covariates"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-subset_covariates'><code>sl3::Lrnr_base$subset_covariates()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="train"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-train'><code>sl3::Lrnr_base$train()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="train_sublearners"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-train_sublearners'><code>sl3::Lrnr_base$train_sublearners()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hte3" data-topic="Lrnr_hte" data-id="check_treatment_type"><a href='../../hte3/html/Lrnr_hte.html#method-Lrnr_hte-check_treatment_type'><code>hte3::Lrnr_hte$check_treatment_type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hte3" data-topic="Lrnr_hte" data-id="get_modifiers"><a href='../../hte3/html/Lrnr_hte.html#method-Lrnr_hte-get_modifiers'><code>hte3::Lrnr_hte$get_modifiers()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hte3" data-topic="Lrnr_hte" data-id="make_metalearner_task"><a href='../../hte3/html/Lrnr_hte.html#method-Lrnr_hte-make_metalearner_task'><code>hte3::Lrnr_hte$make_metalearner_task()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lrnr_cate_EP-new"></a>}}
\if{latex}{\out{\hypertarget{method-Lrnr_cate_EP-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lrnr_cate_EP$new(
  base_learner,
  sieve_num_basis = NULL,
  sieve_interaction_order = 3,
  screen_basis_with_lasso = FALSE,
  treatment_level = NULL,
  control_level = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_learner}}{A \code{\link{sl3}} learner object inheriting from \code{\link[sl3]{Lrnr_base}} that specifies the base supervised learning algorithm used by the meta-learner.}

\item{\code{sieve_num_basis}}{The number of trignometric basis functions to construct the EP-learner sieve space.
This argument is passed as the argument \code{basisN} to the function \code{\link[Sieve]{sieve_preprocess}}.
By default, \code{sieve_num_basis = ceiling((n^(1/3)*d)} where \code{n} denotes the sample size and \code{d} is the number of effect modifiers.}

\item{\code{sieve_interaction_order}}{The maximum interaction degree of tensor-product basis functions in the EP-learner sieve basis. Default is 3.
\code{sieve_interaction_order = 1} corresponds to an additive sieve model and \code{sieve_interaction_order = 2} corresponds to a bi-additive sieve model.
This argument is passed as the argument \code{interaction_order} to the function \code{\link[Sieve]{sieve_preprocess}}.}

\item{\code{screen_basis_with_lasso}}{EXPERIMENTAL. Whether to use a lasso-based DR-learner algorithm to screen sieve basis functions for EP-learner.
There are no theoretical gaurantees for EP-learner with \code{screen_basis_with_lasso = TRUE}. However, this argument may be useful in high dimensional settings.
It also reduces the computational complexity of EP-learner, as the argument \code{sieve_num_basis} does not need to be externally cross-validated.}

\item{\code{...}}{Additional arguments to pass to the initialization function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lrnr_cate_EP-get_pseudo_data"></a>}}
\if{latex}{\out{\hypertarget{method-Lrnr_cate_EP-get_pseudo_data}{}}}
\subsection{Method \code{get_pseudo_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lrnr_cate_EP$get_pseudo_data(
  hte3_task,
  treatment_level = NULL,
  control_level = NULL,
  train = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments to pass to the initialization function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lrnr_cate_EP-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Lrnr_cate_EP-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lrnr_cate_EP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
