% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_cate_T.R
\name{Lrnr_cate_T}
\alias{Lrnr_cate_T}
\title{Lrnr_cate_T Class}
\format{
An R6 class with public methods to initialize the learner, create a regression task, and access the base learner.
}
\description{
This class constructs a T-learner (or S-learner) for estimation of the conditional average treatment effect (CATE)
}
\section{Super classes}{
\code{\link[sl3:Lrnr_base]{sl3::Lrnr_base}} -> \code{hte3::Lrnr_hte} -> \code{Lrnr_cate_T}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Lrnr_cate_T-new}{\code{Lrnr_cate_T$new()}}
\item \href{#method-Lrnr_cate_T-get_pseudo_data}{\code{Lrnr_cate_T$get_pseudo_data()}}
\item \href{#method-Lrnr_cate_T-make_metalearner_task}{\code{Lrnr_cate_T$make_metalearner_task()}}
\item \href{#method-Lrnr_cate_T-clone}{\code{Lrnr_cate_T$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="assert_trained"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-assert_trained'><code>sl3::Lrnr_base$assert_trained()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="base_chain"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-base_chain'><code>sl3::Lrnr_base$base_chain()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="base_predict"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-base_predict'><code>sl3::Lrnr_base$base_predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="base_train"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-base_train'><code>sl3::Lrnr_base$base_train()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="chain"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-chain'><code>sl3::Lrnr_base$chain()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="custom_chain"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-custom_chain'><code>sl3::Lrnr_base$custom_chain()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="get_outcome_range"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-get_outcome_range'><code>sl3::Lrnr_base$get_outcome_range()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="get_outcome_type"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-get_outcome_type'><code>sl3::Lrnr_base$get_outcome_type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="predict"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-predict'><code>sl3::Lrnr_base$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="predict_fold"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-predict_fold'><code>sl3::Lrnr_base$predict_fold()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="print"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-print'><code>sl3::Lrnr_base$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="process_formula"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-process_formula'><code>sl3::Lrnr_base$process_formula()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="reparameterize"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-reparameterize'><code>sl3::Lrnr_base$reparameterize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="retrain"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-retrain'><code>sl3::Lrnr_base$retrain()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="sample"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-sample'><code>sl3::Lrnr_base$sample()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="set_train"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-set_train'><code>sl3::Lrnr_base$set_train()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="subset_covariates"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-subset_covariates'><code>sl3::Lrnr_base$subset_covariates()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="train"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-train'><code>sl3::Lrnr_base$train()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="train_sublearners"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-train_sublearners'><code>sl3::Lrnr_base$train_sublearners()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hte3" data-topic="Lrnr_hte" data-id="check_treatment_type"><a href='../../hte3/html/Lrnr_hte.html#method-Lrnr_hte-check_treatment_type'><code>hte3::Lrnr_hte$check_treatment_type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hte3" data-topic="Lrnr_hte" data-id="get_modifiers"><a href='../../hte3/html/Lrnr_hte.html#method-Lrnr_hte-get_modifiers'><code>hte3::Lrnr_hte$get_modifiers()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lrnr_cate_T-new"></a>}}
\if{latex}{\out{\hypertarget{method-Lrnr_cate_T-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lrnr_cate_T$new(
  base_learner,
  treatment_level = 1,
  control_level = 0,
  stratify_by_treatment = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_learner}}{A \code{\link{sl3}} learner object inheriting from \code{\link[sl3]{Lrnr_base}} that specifies the base supervised learning algorithm used by the meta-learner.}

\item{\code{stratify_by_treatment}}{Logical indicating whether to estimate outcome regression separately in each treatment arm (i.e., T-learner) or pooled across treatment arms (i.e., S-learner).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lrnr_cate_T-get_pseudo_data"></a>}}
\if{latex}{\out{\hypertarget{method-Lrnr_cate_T-get_pseudo_data}{}}}
\subsection{Method \code{get_pseudo_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lrnr_cate_T$get_pseudo_data(hte3_task, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lrnr_cate_T-make_metalearner_task"></a>}}
\if{latex}{\out{\hypertarget{method-Lrnr_cate_T-make_metalearner_task}{}}}
\subsection{Method \code{make_metalearner_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lrnr_cate_T$make_metalearner_task(hte3_task, train = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lrnr_cate_T-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Lrnr_cate_T-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lrnr_cate_T$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
