% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_hte.R
\name{Lrnr_hte}
\alias{Lrnr_hte}
\title{Lrnr_hte Class}
\format{
An R6 class with public methods to initialize the learner, create a regression task, and access the base learner.
}
\description{
This class serves as a general template for constructing meta-learners to estimate heterogeneous treatment effects (HTEs).
}
\keyword{(Used}
\keyword{(untrained)}
\keyword{Access}
\keyword{Base}
\keyword{Construct}
\keyword{Learner}
\keyword{Meta-Learner}
\keyword{Task}
\keyword{This}
\keyword{\code{base_learner}}
\keyword{\code{sl3_Task}}
\keyword{a}
\keyword{algorithm.}
\keyword{base}
\keyword{constructs}
\keyword{during}
\keyword{in}
\keyword{initialization.}
\keyword{internal}
\keyword{internally)}
\keyword{learner}
\keyword{meta-learner}
\keyword{method}
\keyword{object}
\keyword{returns}
\keyword{specified}
\keyword{the}
\keyword{to}
\keyword{train}
\keyword{used}
\section{Super class}{
\code{\link[sl3:Lrnr_base]{sl3::Lrnr_base}} -> \code{Lrnr_hte}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Lrnr_hte-new}{\code{Lrnr_hte$new()}}
\item \href{#method-Lrnr_hte-get_pseudo_data}{\code{Lrnr_hte$get_pseudo_data()}}
\item \href{#method-Lrnr_hte-check_treatment_type}{\code{Lrnr_hte$check_treatment_type()}}
\item \href{#method-Lrnr_hte-get_modifiers}{\code{Lrnr_hte$get_modifiers()}}
\item \href{#method-Lrnr_hte-make_metalearner_task}{\code{Lrnr_hte$make_metalearner_task()}}
\item \href{#method-Lrnr_hte-clone}{\code{Lrnr_hte$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="assert_trained"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-assert_trained'><code>sl3::Lrnr_base$assert_trained()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="base_chain"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-base_chain'><code>sl3::Lrnr_base$base_chain()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="base_predict"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-base_predict'><code>sl3::Lrnr_base$base_predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="base_train"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-base_train'><code>sl3::Lrnr_base$base_train()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="chain"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-chain'><code>sl3::Lrnr_base$chain()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="custom_chain"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-custom_chain'><code>sl3::Lrnr_base$custom_chain()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="get_outcome_range"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-get_outcome_range'><code>sl3::Lrnr_base$get_outcome_range()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="get_outcome_type"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-get_outcome_type'><code>sl3::Lrnr_base$get_outcome_type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="predict"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-predict'><code>sl3::Lrnr_base$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="predict_fold"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-predict_fold'><code>sl3::Lrnr_base$predict_fold()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="print"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-print'><code>sl3::Lrnr_base$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="process_formula"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-process_formula'><code>sl3::Lrnr_base$process_formula()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="reparameterize"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-reparameterize'><code>sl3::Lrnr_base$reparameterize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="retrain"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-retrain'><code>sl3::Lrnr_base$retrain()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="sample"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-sample'><code>sl3::Lrnr_base$sample()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="set_train"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-set_train'><code>sl3::Lrnr_base$set_train()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="subset_covariates"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-subset_covariates'><code>sl3::Lrnr_base$subset_covariates()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="train"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-train'><code>sl3::Lrnr_base$train()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="train_sublearners"><a href='../../sl3/html/Lrnr_base.html#method-Lrnr_base-train_sublearners'><code>sl3::Lrnr_base$train_sublearners()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lrnr_hte-new"></a>}}
\if{latex}{\out{\hypertarget{method-Lrnr_hte-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lrnr_hte$new(
  params,
  base_learner,
  transform_function = NULL,
  pseudo_outcome_type = c("continuous", "binomial", "quasibinomial"),
  pseudo_family = gaussian(),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{A list of parameters for the meta-learning algorithm.}

\item{\code{base_learner}}{A \code{\link{sl3}} learner object inheriting from \code{\link[sl3]{Lrnr_base}} that specifies the base supervised learning algorithm used by the meta-learner.}

\item{\code{transform_function}}{A function to transform the predictions of the base learner. Default is the identity transform.}

\item{\code{pseudo_outcome_type}}{The outcome type of the pseudo-outcome used by the meta-learner. Options are \code{c("continuous", "binomial", "quasibinomial")}. Default is \code{"continuous"}.
For example, the DR-learner, EP-learner, and R-learner of the CATE involve (weighted) least-squares regression using a pseudo-outcome with \code{pseudo_outcome_type = "continuous"}.
The CRATE EP-learner involves performing weighted logistic regression using a pseudo-outcome taking values in [0,1] with \code{pseudo_outcome_type = "quasibinomial"}.}

\item{\code{pseudo_family}}{A \code{\link[stats]{family}} object specifying the loss function (involving pseudo-weights and pseudo-outcomes) used to fit \code{base_learner} in the meta-learner algorithm. Default is \code{gaussian()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lrnr_hte-get_pseudo_data"></a>}}
\if{latex}{\out{\hypertarget{method-Lrnr_hte-get_pseudo_data}{}}}
\subsection{Method \code{get_pseudo_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lrnr_hte$get_pseudo_data(hte3_task, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hte3_task}}{A \code{hte3_Task} object containing the data and necessary information for heterogeneous treatment effect estimation.}

\item{\code{...}}{Additional arguments in \code{params} needed to compute the pseudo-data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing attributes \code{pseudo_outcome} and \code{pseudo_weights}.
(Used internally) Check Compatibility with Treatment Type

This method checks whether this meta-learner is compatible with the \code{treatment} variable type in \code{hte3_Task}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lrnr_hte-check_treatment_type"></a>}}
\if{latex}{\out{\hypertarget{method-Lrnr_hte-check_treatment_type}{}}}
\subsection{Method \code{check_treatment_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lrnr_hte$check_treatment_type(hte3_task)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hte3_task}}{A \code{hte3_Task} object containing the data and necessary information for heterogeneous treatment effect estimation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lrnr_hte-get_modifiers"></a>}}
\if{latex}{\out{\hypertarget{method-Lrnr_hte-get_modifiers}{}}}
\subsection{Method \code{get_modifiers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lrnr_hte$get_modifiers(hte3_task, return_matrix = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lrnr_hte-make_metalearner_task"></a>}}
\if{latex}{\out{\hypertarget{method-Lrnr_hte-make_metalearner_task}{}}}
\subsection{Method \code{make_metalearner_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lrnr_hte$make_metalearner_task(hte3_task, train = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hte3_task}}{A \code{hte3_Task} object containing the data and necessary information for heterogeneous treatment effect estimation.}

\item{\code{train}}{Logical indicating whether to create the task for training or prediction. Default is \code{TRUE}.
If \code{FALSE} then a \code{hte3_Task} object for prediction is returned with covariates being the effect \code{modifiers}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{sl3_Task} object containing the pseudo-data of \code{get_pseudo_data} and \code{outcome_Type}=\code{pseudo_outcome_type}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lrnr_hte-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Lrnr_hte-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lrnr_hte$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
