% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_cross_fitted}
\alias{make_cross_fitted}
\title{Create Cross-Fitted Learner}
\usage{
make_cross_fitted(
  learner,
  calibrate = FALSE,
  cross_validate = inherits(learner, "Stack")
)
}
\arguments{
\item{learner}{The learner to be cross-fitted. This can be a single learner or a stacked learner.}

\item{calibrate}{Logical indicating whether to calibrate the learner (default is \code{FALSE}).}

\item{cross_validate}{Logical indicating whether to perform cross-validation (default is \code{TRUE} if \code{learner} is a stacked learner).}
}
\value{
A cross-fitted version of the input learner.
}
\description{
This function takes a learner and returns a cross-fitted version of it. Cross-fitting involves
fitting the learner to different subsets of the data while using the complementary subsets for validation,
to provide a more robust estimate of model performance.
}
