% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_hte3}
\alias{predict_hte3}
\title{Predict Heterogeneous Treatment Effects using hte3 Learners}
\usage{
predict_hte3(hte3_learner, new_data)
}
\arguments{
\item{hte3_learner}{A trained hte3 learner (\code{Lrnr_sl} object) used to make predictions.}

\item{new_data}{A data frame or data.table containing new data of effect \code{modifiers} for prediction.}
}
\value{
A numeric vector of predicted heterogeneous treatment effects for the provided new data.
}
\description{
Generates treatment effect predictions using a trained hte3 learner.
}
