% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cross_validate_cate}
\alias{cross_validate_cate}
\title{Cross-Validate CATE Models}
\usage{
cross_validate_cate(
  hte_learners,
  hte3_task,
  cv_control = NULL,
  treatment_level = NULL,
  control_level = NULL,
  ...
)
}
\arguments{
\item{hte_learners}{A single \code{Lrnr_hte} meta-learner or a \code{Stack} of \code{Lrnr_hte} meta-learners to cross-validate.}

\item{hte3_task}{An \code{hte3_Task} object containing the data and necessary information for heterogeneous treatment effect estimation.}

\item{cv_control}{A list of control parameters for cross-validation passed to \code{\link[sl3]{Lrnr_sl}}. Default is \code{NULL}.}

\item{treatment_level}{The specific level of the treatment variable to be considered as the treated group. This parameter is crucial for models that differentiate between treated and control groups in estimating treatment effects.}

\item{control_level}{The specific level of the treatment variable to be considered as the control group. Similar to `treatment_level`, this parameter helps in identifying the control group for the purpose of treatment effect estimation.}

\item{...}{Additional arguments to pass to the loss function and other functions.}
}
\value{
A list containing the following elements:
  - `lrnr_sl`: A `Lrnr_sl` object that represents the cross-validated ensemble of CATE learners.
  - `cv_risk`: The cross-validation risk associated with the ensemble, which serves as a measure of the ensemble's predictive performance.
  - `coefficients`: The coefficients derived from the cross-validation process, providing insights into the relative importance of different learners within the ensemble.
}
\description{
Cross-validates a collection of CATE (hte3) learners using the DR-learner loss for the CATE.
}
