% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_m}
\alias{estimate_m}
\title{Helper function for estimation of treatment-averaged outcome regression using \code{sl3}.}
\usage{
estimate_m(
  W,
  Y,
  learner = Lrnr_gam$new(family = "gaussian"),
  weights = NULL,
  cross_fit_and_cv = TRUE,
  return_learner = FALSE,
  folds = 10,
  ...
)
}
\description{
Helper function for estimation of treatment-averaged outcome regression using \code{sl3}.
}
