% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_mu}
\alias{estimate_mu}
\title{Helper function for estimation of outcome regression using \code{sl3}.}
\usage{
estimate_mu(
  W,
  A,
  Y,
  treatment_level,
  learner = Lrnr_gam$new(family = "gaussian"),
  weights = NULL,
  cross_fit_and_cv = TRUE,
  stratified_by_trt = TRUE,
  return_learner = FALSE,
  folds = 10,
  ...
)
}
\description{
Helper function for estimation of outcome regression using \code{sl3}.
}
