% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_pi}
\alias{estimate_pi}
\title{Helper function for estimation of propensity score using \code{sl3}.}
\usage{
estimate_pi(
  W,
  A,
  binomial_learner = Lrnr_gam$new(family = "gaussian"),
  weights = NULL,
  treatment_level = max(A),
  cross_fit_and_cv = TRUE,
  return_learner = FALSE,
  folds = 10,
  ...
)
}
\description{
Helper function for estimation of propensity score using \code{sl3}.
}
