% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_autoML}
\alias{get_autoML}
\title{Get Automated Machine Learning (AutoML) Learner}
\usage{
get_autoML()
}
\value{
A `sl3::Stack` object containing a predefined set of off-the-shelf learners for AutoML.
}
\description{
This function returns a predefined stack of off-the-shelf learning algorithms for Automated Machine Learning (AutoML).
The following base learners are included in the stack:
\itemize{
  \item \code{Lrnr_glmnet}: Elastic Net Regularized Generalized Linear Model.
  \item \code{Lrnr_gam}: Generalized Additive Model.
  \item \code{Lrnr_earth}: Multivariate Adaptive Regression Spline Model.
  \item \code{Lrnr_ranger}: Random Forest Model.
  \item \code{Lrnr_xgboost}: XGBoost Model with different depth configurations.
}
}
