% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lrnr_cate_EP.R
\name{make_ep_stack}
\alias{make_ep_stack}
\title{Create an Ensemble of CATE EP-learners with Varying Sieve Basis Sizes}
\usage{
make_ep_stack(base_learner, hte3_task, treatment_level = 1, control_level = 0)
}
\arguments{
\item{base_learner}{The base learner of \code{Lrnr_cate_EP}.}

\item{hte3_task}{A \code{hte3_Task} object containing the data and necessary information for the heterogeneous treatment effect estimation.}
}
\value{
A stack of CATE EP meta-learners with varying sieve basis sizes.
}
\description{
This function creates an ensemble of CATE EP-learners of varying sieve dimensions for use with cross-validation.
}
